scripts/requirements.txt
# Python dependencies for restaurant analytics data pipeline
# Install with: pip install -r scripts/requirements.txt
# Or from scripts directory: pip install -r requirements.txt

# Data processing
pandas>=2.0.0

# PostgreSQL connection
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Fuzzy string matching for product entity resolution
thefuzz>=0.19.0
python-Levenshtein>=0.21.0  # Optional but recommended for faster fuzzy matching

# Timezone handling
pytz>=2023.3

# Environment variable management
python-dotenv>=1.0.0